Add fractional temperature steps and enhanced climate display
Enables precise 0.5°C temperature control and improves climate display with current temperature tracking, dual-circle layout showing current vs setpoint, arc markers, and crash fixes for climate API calls.

Key improvements:
- Fractional rotary_step_width support (0.5°C increments) via x10 scaling
- Subscribe to current_temperature attribute from Home Assistant
- Dual-circle display: left (current temp) vs right (setpoint)
- Arc with color-coded gap and position markers
- Fixed null pointer crashes in turnClimateOn/Off (missing init calls)
- Display refresh flags for real-time updates
- Black background for better contrast
- Custom fireplace icon (70x70 RGB565)

+++ components/shys_m5_dial/shys_m5_dial.h
@@ -28,7 +28,7 @@ namespace esphome
     protected:
       int timeToScreenOff = 30000;
       int longPressMs = 1200;
-      int rotaryStepWidth = 10;
+      float rotaryStepWidth = 10.0;
       uint16_t displayRefeshPause = 700;
 
       int apiSendDelay = 1000; // Verzögerung nach Wert-Änderung (um nicht jeden Wert beim drehen des Rades zu senden)
@@ -183,8 +183,8 @@ namespace esphome
           this->apiSendLock = delayInMs;
       }
 
-      void setRotaryStepWidth(int value){
-        ESP_LOGI("DEVICE", "setRotaryStepWidth %i", value);
+      void setRotaryStepWidth(float value){
+        ESP_LOGI("DEVICE", "setRotaryStepWidth %.2f", value);
         this->rotaryStepWidth = value;
       }
